Add indicator to zsh prompt if background jobs
I often sleep (`^Z`) my vim session to do something on the command line
and then restore (`fg`) when I'm done. However, I often forget if I have
a backgrounded vim and try to restore it.

This commit adds a little yellow dot to the prompt if there are any
background jobs.